requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
bcrypt==4.1.1
python-dotenv==1.0.0
Werkzeug==3.0.1
WTForms==3.1.1
email-validator==2.1.0
Pillow==10.1.0
pytest==7.4.3
pytest-flask==1.3.0
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
icalendar==5.0.11
bleach==6.1.0
python-dateutil==2.8.2
pytz==2023.3
